+++ test/clj_ml/classifiers_test.clj
@@ -1,5 +1,5 @@
-(ns clj-ml.classifiers-test
-  (:use [clj-ml classifiers data] :reload-all)
+(ns clj-ml-dev.classifiers-test
+  (:use [clj-ml-dev classifiers data] :reload-all)
   (:use clojure.test midje.sweet))
 
 
@@ -18,36 +18,36 @@
 
 (deftest train-classifier-c45
   (let [c (make-classifier :decision-tree :c45)
-        ds (clj-ml.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])]
-    (clj-ml.data/dataset-set-class ds 2)
+        ds (clj-ml-dev.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])]
+    (clj-ml-dev.data/dataset-set-class ds 2)
     (classifier-train c ds)
     (is true)))
 
 (deftest make-classifier-bayes
   (fact
-    (let [c (clj-ml.classifiers/make-classifier :bayes :naive {:kernel-estimator true :old-format true})
+    (let [c (clj-ml-dev.classifiers/make-classifier :bayes :naive {:kernel-estimator true :old-format true})
           opts (vec (.getOptions c))]
       opts => (contains ["-K" "-O"]))))
 
 (deftest make-classifier-bayes-updateable
-  (let [c (clj-ml.classifiers/make-classifier :bayes :naive {:updateable true})]
+  (let [c (clj-ml-dev.classifiers/make-classifier :bayes :naive {:updateable true})]
     (is (= (class c)
            weka.classifiers.bayes.NaiveBayesUpdateable))))
 
 (deftest train-classifier-bayes
-  (let [c (clj-ml.classifiers/make-classifier :bayes :naive {:kernel-estimator true :old-format true})
-        ds (clj-ml.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])]
-    (clj-ml.data/dataset-set-class ds 2)
+  (let [c (clj-ml-dev.classifiers/make-classifier :bayes :naive {:kernel-estimator true :old-format true})
+        ds (clj-ml-dev.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])]
+    (clj-ml-dev.data/dataset-set-class ds 2)
     (classifier-train c ds)
     (is true)))
 
 
 (deftest classifier-evaluate-dataset
   (let [c   (make-classifier :decision-tree :c45)
-        ds  (clj-ml.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])
-        tds (clj-ml.data/make-dataset "test" [:a :b {:c [:m :n]}] [[4 1 :n] [4 5 :m]])
-        _   (clj-ml.data/dataset-set-class ds 2)
-        _   (clj-ml.data/dataset-set-class tds 2)
+        ds  (clj-ml-dev.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])
+        tds (clj-ml-dev.data/make-dataset "test" [:a :b {:c [:m :n]}] [[4 1 :n] [4 5 :m]])
+        _   (clj-ml-dev.data/dataset-set-class ds 2)
+        _   (clj-ml-dev.data/dataset-set-class tds 2)
         _   (classifier-train c ds)
         res (classifier-evaluate c :dataset ds tds)]
     (is (= 26 (count (keys res))))))
@@ -60,15 +60,15 @@
 
 (deftest classifier-evaluate-cross-validation
   (let [c (make-classifier :decision-tree :c45)
-        ds (clj-ml.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])
-        _  (clj-ml.data/dataset-set-class ds 2)
+        ds (clj-ml-dev.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])
+        _  (clj-ml-dev.data/dataset-set-class ds 2)
         _  (classifier-train c ds)
         res (classifier-evaluate c :cross-validation ds 2)]
     (is (= 26 (count (keys res))))))
 
 (deftest update-updateable-classifier
-  (let [c (clj-ml.classifiers/make-classifier :bayes :naive {:updateable true})
-        ds (clj-ml.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])
+  (let [c (clj-ml-dev.classifiers/make-classifier :bayes :naive {:updateable true})
+        ds (clj-ml-dev.data/make-dataset "test" [:a :b {:c [:m :n]}] [[1 2 :m] [4 5 :m]])
         _  (dataset-set-class ds 2)
         inst (make-instance ds {:a 56 :b 45 :c :m})]
     (classifier-train  c ds)